Assign value without any control statement
Given four integers ‘a’, ‘b’, ‘y’ and ‘x’, where ‘x’ can only be either zero or one. Your task is as follows: 
 

If ‘x’ is zero assign value ‘a’ to the variable ‘y’
If ‘x’ is one assign value ‘b’ to the variable ‘y’.
It is not allowed to use any conditional operator (including the ternary operator).